Dotfiles

+++ termux-setup.sh
@@ -11,18 +11,24 @@ pkg install -y nano git openssh python python-dev python2 coreutils proot
 if [[ ! -d /data/data/com.termux/files/home/storage/downloads/ssh ]]; then
   echo "SSH key directory not present"
 else
-  cp -R /data/data/com.termux/files/home/storage/downloads/ssh ~/.ssh
+  cp -R /data/data/com.termux/files/home/storage/downloads/ssh/. ~/.ssh/
 fi
 
 if [[ -d /data/data/com.termux/files/home/storage/downloads/src ]]; then
-  ln -s /data/data/com.termux/files/home/storage/downloads/src ~/src
+  if [[! -e "~/src" ]]; then
+    ln -s /data/data/com.termux/files/home/storage/downloads/src ~/src
+  fi
 fi
 
-if [[ -f "~/.bashrc" ]]; then
-  rm -f ~/.bashrc.bak
-  mv "~/.bashrc" ~/.bashrc.bak
-fi
+# alrighty, let's sort out our dotfiles
+git clone https://github.com/grahamgilbert/termux-dotfiles.git "$TMPDIR/termux-dotfiles"
 
-echo "alias ipaddress=\"ifconfig arc0 | awk '/inet /{print $2}'\"" >> ~/.bashrc
-termux-chroot
-sshd
+for f in $TMPDIR/termux-dotfiles/*
+do
+  if [ "$f" != "README.md" ]; then
+    cp -f $TMPDIR/termux-dotfiles/$f "~/.${f}"
+  fi
+done
+
+rm -rf $TMPDIR/termux-dotfiles
+source ~/.bashrc
